requirements.txt
flask==2.3.3
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
werkzeug==2.3.7
tensorflow-datasets==4.9.2
requests==2.31.0
tqdm==4.66.1
h5py==3.9.0
